style: clear the clippy backlog outside the engine's own gate
CI runs cargo fmt --check and the workspace tests, so these had accumulated
unnoticed: needless range loops, &mut Vec where a slice will do, as_ref().map
deref chains, and a handful of doc-list indentation warnings.

Mechanical, and kept apart from the fixes beside them so neither has to be read
through the other.

+++ crates/aurora-interp/src/lib.rs
@@ -790,8 +790,8 @@ impl<'a> Interp<'a> {
                 Ok(Value::Unit)
             }
             "fill_rect_alpha" => {
-                let (x, y) = (int_arg(&argv, 0) as i64, int_arg(&argv, 1) as i64);
-                let (w, h) = (int_arg(&argv, 2) as i64, int_arg(&argv, 3) as i64);
+                let (x, y) = (int_arg(&argv, 0), int_arg(&argv, 1));
+                let (w, h) = (int_arg(&argv, 2), int_arg(&argv, 3));
                 let c = color_arg(&argv, 4);
                 let a = int_arg(&argv, 7).clamp(0, 255) as u8;
                 if let Some(fb) = &mut self.gfx {
